% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{createReport}
\alias{createReport}
\title{Create rmarkdown report}
\usage{
createReport(output_dir = getwd())
}
\arguments{
\item{output_dir}{optional out directory to write rmarkdown to}
}
\value{
none, creates rmarkdown report
}
\description{
Create rmarkdown report
}
