% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readNoaa}
\alias{readNoaa}
\title{Read data from NOAA Real-Time Meteorological Observation Network}
\usage{

  readNoaa(con = "https://www.glerl.noaa.gov/metdata/chi/2019/20190421.04t.txt")
}
\arguments{
\item{con}{url}
}
\value{
data.frame
}
\description{
Read data from NOAA Real-Time Meteorological Observation Network
}
